fileset: support multiple names in a new `OpenMultiple` method
The primary use case is to provide a notion of precedence in
configuration files. This will allow us to have, for example, a "base"
configuration and then an "override" configuration. In the simple
example:

    /etc/base.toml
    /etc/override.toml

If we call `OpenMultiple({"base", "override"}, {"/etc"})` then we'll
first apply the configuration in `/etc/base.toml` and then the
configuration in `/etc/override.toml` which will allow us to override
parts of the base configuration in a consistent and predictable way.