marshal: do not invoke nil TextMarshalers
This change ensures that we emit a nil when marshaling a nil value
implementing encoding.TextMarshaler, instead of calling MarshalText() on
nil, which a lot of implementations aren't prepared to deal with.

Because of this, and the recent addition of regexp.Regexp.MarshalText(),
we also need to adjust the marshaled value for regexps: instead of
emitting an empty string, we instead treat nil regexps as nil.